+++ tests/test_atoms.py
@@ -427,11 +427,11 @@ def test_eqm_bond_distance():
 
     h2_atoms = Atoms([Atom('H'), Atom('H')])
 
-    assert np.isclose(h2_atoms.eqm_distance(0, 0), 0.)
+    assert np.isclose(h2_atoms.eqm_bond_distance(0, 0), 0.)
 
     assert np.isclose(h2_atoms.distance(0, 1), 0.)
-    assert np.isclose(h2_atoms.eqm_distance(0, 1), 0.7, atol=0.1)
+    assert np.isclose(h2_atoms.eqm_bond_distance(0, 1), 0.7, atol=0.1)
 
     # Cannot determine the distance between atoms not present in the set
     with pytest.raises(ValueError):
-        _ = h2_atoms.eqm_distance(0, 2)
+        _ = h2_atoms.eqm_bond_distance(0, 2)